requirements.txt
Django==2.0.3
django-bootstrap4==0.0.6
django-ckeditor==5.4.0
django-js-asset==1.1.0
Pillow==5.1.0
pytz==2018.3
